requirements.txt
apache-flink==1.19.0
apache_flink_libraries==1.19.0
# psycopg2==2.9.9
psycopg2_binary==2.9.10
python-dotenv==1.2.1
